extract rendering orchestrators from ReportScreen
