+++ tests/unit/AgentZeroStatusContextTest.ts
@@ -46,7 +46,7 @@ function simulateReasoning(data: {reasoning: string; agentZeroRequestID: string;
 const reportID = '123';
 
 function wrapper({children}: {children: React.ReactNode}) {
-    return React.createElement(AgentZeroStatusProvider, {reportID, chatType: undefined}, children);
+    return React.createElement(AgentZeroStatusProvider, {reportID}, children);
 }
 
 describe('AgentZeroStatusContext', () => {
